+++ src/runtime/arch.h
@@ -26,8 +26,8 @@ extern boolean arch_pseudo_atomic_atomic(os_context_t*);
 extern void arch_set_pseudo_atomic_interrupted(os_context_t*);
 extern os_vm_address_t arch_get_bad_addr(int, siginfo_t*, os_context_t*);
 extern unsigned char *arch_internal_error_arguments(os_context_t*);
-extern unsigned long arch_install_breakpoint(void *pc);
-extern void arch_remove_breakpoint(void *pc, unsigned long orig_inst);
+extern unsigned int arch_install_breakpoint(void *pc);
+extern void arch_remove_breakpoint(void *pc, unsigned int orig_inst);
 extern void arch_install_interrupt_handlers(void);
 extern void arch_do_displaced_inst(os_context_t *context,
                                    unsigned int orig_inst);
